chore(spec): added remaining benchmark already present in spec

+++ tfhe-benchmark/benches/integer/bench.rs
@@ -151,7 +151,7 @@ fn bench_server_key_binary_function_clean_inputs<F>(
             &benchmark_spec,
             display_name,
             &OperatorType::Atomic,
-            bit_size as u64,
+            u64::from(bit_size),
             vec![param.message_modulus().0.ilog2(); num_block],
         );
     }
@@ -224,7 +224,7 @@ fn bench_server_key_unary_function_dirty_inputs<F>(
             &benchmark_spec,
             display_name,
             &OperatorType::Atomic,
-            bit_size as u64,
+            u64::from(bit_size),
             vec![param.message_modulus().0.ilog2(); num_block],
         );
     }
@@ -331,7 +331,7 @@ fn bench_server_key_unary_function_clean_inputs<F>(
             &benchmark_spec,
             display_name,
             &OperatorType::Atomic,
-            bit_size as u64,
+            u64::from(bit_size),
             vec![param.message_modulus().0.ilog2(); num_block],
         );
     }
@@ -347,7 +347,7 @@ fn bench_server_key_binary_scalar_function_clean_inputs<F, G>(
     rng_func: G,
 ) where
     F: Fn(&ServerKey, &RadixCiphertext, ScalarType) + Sync,
-    G: Fn(&mut ThreadRng, usize) -> ScalarType,
+    G: Fn(&mut ThreadRng, u32) -> ScalarType,
 {
     let mut bench_group = c.benchmark_group(integer_op.to_string());
     bench_group
@@ -356,12 +356,12 @@ fn bench_server_key_binary_scalar_function_clean_inputs<F, G>(
     let mut rng = rand::thread_rng();
 
     for (param, num_block, bit_size) in ParamsAndNumBlocksIter::default() {
-        if bit_size > ScalarType::BITS as usize {
+        if bit_size > ScalarType::BITS {
             break;
         }
         let param_name = param.name();
 
-        let max_value_for_bit_size = ScalarType::MAX >> (ScalarType::BITS as usize - bit_size);
+        let max_value_for_bit_size = ScalarType::MAX >> (ScalarType::BITS - bit_size);
 
         // Preserve the historical id suffix: latency kept `_bits_scalar_{bit_size}`,
         // throughput never had it.
@@ -460,7 +460,7 @@ fn bench_server_key_binary_scalar_function_clean_inputs<F, G>(
             &benchmark_spec,
             display_name,
             &OperatorType::Atomic,
-            bit_size as u64,
+            u64::from(bit_size),
             vec![param.message_modulus().0.ilog2(); num_block],
         );
     }
@@ -469,16 +469,16 @@ fn bench_server_key_binary_scalar_function_clean_inputs<F, G>(
 }
 
 // Functions used to apply different way of selecting a scalar based on the context.
-fn default_scalar(rng: &mut ThreadRng, _clear_bit_size: usize) -> ScalarType {
+fn default_scalar(rng: &mut ThreadRng, _clear_bit_size: u32) -> ScalarType {
     gen_random_u256(rng)
 }
 
-fn shift_scalar(_rng: &mut ThreadRng, _clear_bit_size: usize) -> ScalarType {
+fn shift_scalar(_rng: &mut ThreadRng, _clear_bit_size: u32) -> ScalarType {
     // Shifting by one is the worst case scenario.
     ScalarType::ONE
 }
 
-fn mul_scalar(rng: &mut ThreadRng, _clear_bit_size: usize) -> ScalarType {
+fn mul_scalar(rng: &mut ThreadRng, _clear_bit_size: u32) -> ScalarType {
     loop {
         let scalar = gen_random_u256(rng);
         // If scalar is power of two, it is just a shit, which is an happy path.
@@ -488,10 +488,10 @@ fn mul_scalar(rng: &mut ThreadRng, _clear_bit_size: usize) -> ScalarType {
     }
 }
 
-fn div_scalar(rng: &mut ThreadRng, clear_bit_size: usize) -> ScalarType {
+fn div_scalar(rng: &mut ThreadRng, clear_bit_size: u32) -> ScalarType {
     loop {
         let scalar = gen_random_u256(rng);
-        let max_for_bit_size = ScalarType::MAX >> (ScalarType::BITS as usize - clear_bit_size);
+        let max_for_bit_size = ScalarType::MAX >> (ScalarType::BITS - clear_bit_size);
         let scalar = scalar & max_for_bit_size;
         if scalar != ScalarType::ZERO {
             return scalar;
@@ -605,7 +605,7 @@ fn if_then_else_parallelized(c: &mut Criterion) {
             &benchmark_spec,
             display_name,
             &OperatorType::Atomic,
-            bit_size as u64,
+            u64::from(bit_size),
             vec![param.message_modulus().0.ilog2(); num_block],
         );
     }
@@ -719,7 +719,7 @@ fn flip_parallelized(c: &mut Criterion) {
             &benchmark_spec,
             display_name,
             &OperatorType::Atomic,
-            bit_size as u64,
+            u64::from(bit_size),
             vec![param.message_modulus().0.ilog2(); num_block],
         );
     }
@@ -739,7 +739,7 @@ fn ciphertexts_sum_parallelized(c: &mut Criterion) {
 
     for (param, num_block, bit_size) in ParamsAndNumBlocksIter::default() {
         let param_name = param.name();
-        let max_for_bit_size = ScalarType::MAX >> (ScalarType::BITS as usize - bit_size);
+        let max_for_bit_size = ScalarType::MAX >> (ScalarType::BITS - bit_size);
 
         for len in [5, 10, 20] {
             let bits = PrecisionTag::Bits(bit_size);
@@ -780,7 +780,7 @@ fn ciphertexts_sum_parallelized(c: &mut Criterion) {
                 BenchmarkType::Throughput => {
                     let (cks, sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
 
-                    let nb_ctxt = bit_size.div_ceil(param.message_modulus().0.ilog2() as usize);
+                    let nb_ctxt = bit_size.div_ceil(param.message_modulus().0.ilog2()) as usize;
                     let cks = RadixClientKey::from((cks, nb_ctxt));
 
                     bench_group
@@ -845,7 +845,7 @@ fn ciphertexts_sum_parallelized(c: &mut Criterion) {
                 &benchmark_spec,
                 display_name,
                 &OperatorType::Atomic,
-                bit_size as u64,
+                u64::from(bit_size),
                 vec![param.message_modulus().0.ilog2(); num_block],
             );
         }
@@ -1374,7 +1374,7 @@ mod cuda {
                 &benchmark_spec,
                 display_name,
                 &OperatorType::Atomic,
-                bit_size as u64,
+                u64::from(bit_size),
                 vec![param.message_modulus().0.ilog2(); num_block],
             );
         }
@@ -1523,7 +1523,7 @@ mod cuda {
                 &benchmark_spec,
                 display_name,
                 &OperatorType::Atomic,
-                bit_size as u64,
+                u64::from(bit_size),
                 vec![param.message_modulus().0.ilog2(); num_block],
             );
         }
@@ -1541,19 +1541,19 @@ mod cuda {
     ) where
         F: Fn(&CudaServerKey, &mut CudaUnsignedRadixCiphertext, ScalarType, &CudaStreams) + Sync,
         G: Fn(&ServerKey, &mut RadixCiphertext, ScalarType) + Sync,
-        H: Fn(&mut ThreadRng, usize) -> ScalarType,
+        H: Fn(&mut ThreadRng, u32) -> ScalarType,
     {
         let mut bench_group = c.benchmark_group(integer_op.to_string());
         let mut rng = rand::thread_rng();
 
         for (param, num_block, bit_size) in ParamsAndNumBlocksIter::default() {
-            if bit_size > ScalarType::BITS as usize {
+            if bit_size > ScalarType::BITS {
                 break;
             }
 
             let param_name = param.name();
 
-            let max_value_for_bit_size = ScalarType::MAX >> (ScalarType::BITS as usize - bit_size);
+            let max_value_for_bit_size = ScalarType::MAX >> (ScalarType::BITS - bit_size);
 
             // Preserve the historical `_bits_scalar_{bit_size}` id suffix (present on
             // both latency and throughput for the GPU scalar benches).
@@ -1666,7 +1666,7 @@ mod cuda {
                 &benchmark_spec,
                 display_name,
                 &OperatorType::Atomic,
-                bit_size as u64,
+                u64::from(bit_size),
                 vec![param.message_modulus().0.ilog2(); num_block],
             );
         }
@@ -1683,7 +1683,7 @@ mod cuda {
         let mut rng = rand::thread_rng();
 
         for (param, num_block, bit_size) in ParamsAndNumBlocksIter::default() {
-            if bit_size > ScalarType::BITS as usize {
+            if bit_size > ScalarType::BITS {
                 break;
             }
 
@@ -1822,7 +1822,7 @@ mod cuda {
                 &benchmark_spec,
                 "if_then_else",
                 &OperatorType::Atomic,
-                bit_size as u64,
+                u64::from(bit_size),
                 vec![param.message_modulus().0.ilog2(); num_block],
             );
         }
@@ -2667,7 +2667,7 @@ mod cuda {
                     target_num_blocks * param.message_modulus().0.ilog2() as usize;
                 let conversion = PrecisionTag::Conversion {
                     from: bit_size,
-                    to: target_bit_size,
+                    to: target_bit_size as u32,
                 };
                 let benchmark_spec = BenchmarkSpec::new_integer_ops(
                     IntegerOpBySign::Unsigned(integer_op),
@@ -2699,7 +2699,7 @@ mod cuda {
                     &benchmark_spec,
                     display_name,
                     &OperatorType::Atomic,
-                    bit_size as u64,
+                    u64::from(bit_size),
                     vec![param.message_modulus().0.ilog2(); num_blocks],
                 );
             }
@@ -2762,12 +2762,12 @@ mod hpu {
         let mut rng = rand::thread_rng();
 
         for (param, num_block, bit_size) in ParamsAndNumBlocksIter::default() {
-            if bit_size > ScalarType::BITS as usize {
+            if bit_size > ScalarType::BITS {
                 break;
             }
             let param_name = param.name();
 
-            let max_value_for_bit_size = ScalarType::MAX >> (ScalarType::BITS as usize - bit_size);
+            let max_value_for_bit_size = ScalarType::MAX >> (ScalarType::BITS - bit_size);
 
             let bits = PrecisionTag::Bits(bit_size);
             let benchmark_spec = BenchmarkSpec::new_integer_ops(
@@ -2807,7 +2807,7 @@ mod hpu {
 
                                     let clear = rng
                                         .gen_range(0..u128::cast_from(max_value_for_bit_size))
-                                        & if bw < u128::BITS as usize {
+                                        & if bw < u128::BITS {
                                             (1_u128 << bw) - 1
                                         } else {
                                             !0_u128
@@ -2888,7 +2888,7 @@ mod hpu {
 
                                         let clear = rng
                                             .gen_range(0..u128::cast_from(max_value_for_bit_size))
-                                            & if bw < u128::BITS as usize {
+                                            & if bw < u128::BITS {
                                                 (1_u128 << bw) - 1
                                             } else {
                                                 !0_u128
@@ -2942,7 +2942,7 @@ mod hpu {
                 &benchmark_spec,
                 display_name,
                 &OperatorType::Atomic,
-                bit_size as u64,
+                u64::from(bit_size),
                 vec![param.message_modulus().0.ilog2(); num_block],
             );
         }
@@ -3437,7 +3437,7 @@ fn bench_server_key_cast_function<F>(
             let target_bit_size = target_num_blocks * param.message_modulus().0.ilog2() as usize;
             let conversion = PrecisionTag::Conversion {
                 from: bit_size,
-                to: target_bit_size,
+                to: target_bit_size as u32,
             };
             let benchmark_spec = BenchmarkSpec::new_integer_ops(
                 IntegerOpBySign::Unsigned(integer_op),
@@ -3465,7 +3465,7 @@ fn bench_server_key_cast_function<F>(
                 &benchmark_spec,
                 display_name,
                 &OperatorType::Atomic,
-                bit_size as u64,
+                u64::from(bit_size),
                 vec![param.message_modulus().0.ilog2(); num_blocks],
             );
         }
